src/utils/common_utils.py
import os
import sys

import yaml

from src.exception import TSException
from src.logger import logging


def read_yaml_file(file_path: str) -> dict:
    try:
        with open(file_path, "rb") as yaml_file:
            return yaml.safe_load(yaml_file)

    except Exception as e:
        raise TSException(e, sys) from e