requirements.txt
from_root
boto3
pandas
pyarrow
fastparquet
python-dotenv
PyYAML

-e .